2020 day10 p2 optimizations
Not really noticeable when timing as the cache dominates
the running time so much, but I found it fun to optimize
the vector.erase(begin, i) to reverse it and do effectively
pop_back() but via erase(end - n, end).  This minor
optimization prevents the 'remaining' vector from having
to shuffle values around when new permutation combinations
are found.